Separate meta API and utility functions from LinkBackEnd
